refactor(2fa): moves files into new settings app; updates references
