+++ src/v2/Apps/Settings/Routes/EditSettings/Components/SettingsEditSettingsTwoFactor/TwoFactorAuthentication/Components/BackupSecondFactor/__tests__/index.jest.tsx
@@ -1,7 +1,7 @@
 import { Button } from "@artsy/palette"
 import { mount } from "enzyme"
 
-import { BackupSecondFactor } from "../"
+import { BackupSecondFactor } from ".."
 import {
   AppEnabledWithBackupCodesQueryResponse,
   DisabledQueryResponse,
